authentication

+++ graphql/schemas/user/resolver.js
@@ -1,20 +1,33 @@
 const axios = require('axios')
 const { userApi } = require('../../../config/api')
 const jwt = require('jsonwebtoken')
+const { AuthenticationError } = require('apollo-server-express')
 
 module.exports.resolver = {
   Query: {
-    users: async () => {
-      const { data } = await axios.get(`${userApi}/user`)
-      return data
+    users: async (parent, args, context) => {
+      if (context.data) {
+        const { data } = await axios.get(`${userApi}/user`)
+        return data
+      } else {
+        throw new AuthenticationError('Must Authenticate')
+      }
     },
     user: async (parent, { _id }) => {
-      const { data } = await axios.get(`${userApi}/user/${_id}`)
-      return data
+      if (context.data) {
+        const { data } = await axios.get(`${userApi}/user/${_id}`)
+        return data
+      } else {
+        throw new AuthenticationError('Must Authenticate')
+      }
     },
     profile: async (parent, args, context) => {
-      const { data } = await axios.get(`${userApi}/user/${context.data}`)
-      return data
+      if (context.data) {
+        const { data } = await axios.get(`${userApi}/user/${context.data}`)
+        return data
+      } else {
+        throw new AuthenticationError('Must Authenticate')
+      }
     },
     login: async (parent, { username, password }) => {
       const { data } = await axios.post(`${userApi}/user/authenticate`, {
@@ -25,8 +38,12 @@ module.exports.resolver = {
         : ''
     },
     search: async (parent, { skill }) => {
-      const { data } = await axios.post(`${userApi}/user/search`, { skill })
-      return data
+      if (context.data) {
+        const { data } = await axios.post(`${userApi}/user/search`, { skill })
+        return data
+      } else {
+        throw new AuthenticationError('Must Authenticate')
+      }
     }
   },
   Mutation: {
@@ -35,14 +52,22 @@ module.exports.resolver = {
       return data
     },
     deleteUser: async (parent, args, context) => {
-      const { data } = await axios.delete(`${userApi}/user/${context.data}`)
-      return data
+      if (context.data) {
+        const { data } = await axios.delete(`${userApi}/user/${context.data}`)
+        return data
+      } else {
+        throw new AuthenticationError('Must Authenticate')
+      }
     },
     updateUser: async (parent, { user }, context) => {
-      const { data } = await axios.put(`${userApi}/user/${context.data}`, {
-        user
-      })
-      return data
+      if (context.data) {
+        const { data } = await axios.put(`${userApi}/user/${context.data}`, {
+          user
+        })
+        return data
+      } else {
+        throw new AuthenticationError('Must Authenticate')
+      }
     }
   }
 }
